# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "船舶能效分析平台 API",
    "description": "整合 Meta / Identity / Vessel / Data 四个微服务的 OpenAPI 文档。\n\n每条路径通过 path-level `servers` 字段标明所属服务：\n\n| 服务 | 地址 |\n|------|------|\n| Meta 元数据 | http://localhost:8000 |\n| Identity 身份认证 | http://localhost:8001 |\n| Vessel 船舶管理 | http://localhost:8002 |\n| Data 遥测数据 | http://localhost:8003 |",
    "version": "0.1.0"
  },
  "paths": {
...
    "/company": {
      "get": {
        "tags": [
          "公司"
        ],
        "summary": "获取所有公司",
        "description": "返回系统内所有公司列表，无分页。",
        "operationId": "get_companies_company_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponseModel_list_CompanySchema__"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "公司"
        ],
        "summary": "创建公司",
        "description": "创建新公司。公司名称全局唯一。",
        "operationId": "create_company_company_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentityCompanyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponseModel_CompanySchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityHTTPValidationError"
                }
              }
            }
          }
        }
      },
      "servers": [
        {
          "url": "http://localhost:8001",
          "description": "Identity 身份认证服务"
        }
      ]
    },
    "/company/{company_id}": {
      "get": {
        "tags": [
          "公司"
        ],
        "summary": "获取单个公司详情",
        "description": "按 ID 获取公司信息。不存在时返回 404。",
        "operationId": "get_company_company__company_id__get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponseModel_CompanySchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityHTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "公司"
        ],
        "summary": "更新公司信息",
        "description": "部分更新公司信息，仅传入需要修改的字段。",
        "operationId": "update_company_company__company_id__put",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Company Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentityCompanyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponseModel_CompanySchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityHTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "公司"
        ],
        "summary": "删除公司",
        "description": "硬删除公司记录。请确认该公司下无用户，否则会导致用户的 company_id 悬空。",
        "operationId": "delete_company_company__company_id__delete",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponseModel_CompanySchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityHTTPValidationError"
                }
              }
            }
          }
        }
      },
      "servers": [
        {
          "url": "http://localhost:8001",
          "description": "Identity 身份认证服务"
        }
      ]
    },
    "/user": {
      "get": {
        "tags": [
          "用户"
        ],
        "summary": "获取用户列表",
        "description": "获取未禁用的用户列表，支持按用户名模糊搜索和公司过滤，支持分页。",
        "operationId": "get_user_list_user_get",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "用户名(模糊搜索)",
              "title": "Name"
            },
            "description": "用户名(模糊搜索)"
          },
          {
            "name": "company_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "公司ID",
              "title": "Company Id"
            },
            "description": "公司ID"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "偏移量",
              "default": 0,
              "title": "Offset"
            },
            "description": "偏移量"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "每页数量",
              "default": 10,
              "title": "Limit"
            },
            "description": "每页数量"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponseModel_list_UserSchema__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityHTTPValidationError"
                }
              }
            }
          }
        }
      },
      "servers": [
        {
          "url": "http://localhost:8001",
          "description": "Identity 身份认证服务"
        }
      ]
    },
    "/user/register": {
      "post": {
        "tags": [
          "用户"
        ],
        "summary": "注册用户",
        "description": "创建新用户，密码自动加密存储，响应中不包含密码字段。",
        "operationId": "register_user_user_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentityUserRegisterData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponseModel_UserSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityHTTPValidationError"
                }
              }
            }
          }
        }
      },
      "servers": [
        {
          "url": "http://localhost:8001",
          "description": "Identity 身份认证服务"
        }
      ]
    },
    "/user/login": {
      "post": {
        "tags": [
          "用户"
        ],
        "summary": "用户登录",
        "description": "验证用户名和密码，成功后返回用户信息及 JWT token。token 有效期由服务配置决定。",
        "operationId": "login_user_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentityUserLoginData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponseModel_UserWithToken_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityHTTPValidationError"
                }
              }
            }
          }
        }
      },
      "servers": [
        {
          "url": "http://localhost:8001",
          "description": "Identity 身份认证服务"
        }
      ]
    },
    "/user/{user_id}": {
      "get": {
        "tags": [
          "用户"
        ],
        "summary": "获取用户信息",
        "description": "按 ID 获取用户信息。被禁用的用户等同于不存在，返回 404。",
        "operationId": "get_user_user__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponseModel_UserSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityHTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "用户"
        ],
        "summary": "更新用户信息",
        "description": "目前仅支持更新手机号。",
        "operationId": "update_user_user__user_id__put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentityUserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponseModel_UserSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityHTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "用户"
        ],
        "summary": "删除用户",
        "description": "软删除：disabled=true，数据保留。被禁用用户无法登录，GET 返回 404。",
        "operationId": "delete_user_user__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponseModel_UserSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityHTTPValidationError"
                }
              }
            }
          }
        }
      },
      "servers": [
        {
          "url": "http://localhost:8001",
          "description": "Identity 身份认证服务"
        }
      ]
    },
...
  },
  "components": {
    "schemas": {
...
      "IdentityCompanyCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "contact_person": {
            "type": "string",
            "title": "Contact Person"
          },
          "contact_phone": {
            "type": "string",
            "title": "Contact Phone"
          },
          "contact_email": {
            "type": "string",
            "title": "Contact Email"
          }
        },
        "type": "object",
        "required": [
          "name",
          "address",
          "contact_person",
          "contact_phone",
          "contact_email"
        ],
        "title": "CompanyCreate",
        "examples": [
          {
            "address": "上海市浦东新区张江高科技园区",
            "contact_email": "contact@shipping.com",
            "contact_person": "张三",
            "contact_phone": "13800000001",
            "name": "远洋航运有限公司"
          }
        ]
      },
      "IdentityCompanySchema": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "contact_person": {
            "type": "string",
            "title": "Contact Person"
          },
          "contact_phone": {
            "type": "string",
            "title": "Contact Phone"
          },
          "contact_email": {
            "type": "string",
            "title": "Contact Email"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "address",
          "contact_person",
          "contact_phone",
          "contact_email"
        ],
        "title": "CompanySchema",
        "examples": [
          {
            "address": "上海市浦东新区张江高科技园区",
            "contact_email": "contact@shipping.com",
            "contact_person": "张三",
            "contact_phone": "13800000001",
            "id": 1,
            "name": "远洋航运有限公司"
          }
        ]
      },
      "IdentityCompanyUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "contact_person": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Person"
          },
          "contact_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Phone"
          },
          "contact_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          }
        },
        "type": "object",
        "title": "CompanyUpdate",
        "examples": [
          {
            "address": "北京市朝阳区建国路88号"
          }
        ]
      },
      "IdentityHTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/IdentityValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IdentityResponseModel_CompanySchema_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "data": {
            "$ref": "#/components/schemas/IdentityCompanySchema"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ResponseModel[CompanySchema]"
      },
      "IdentityResponseModel_UserSchema_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "data": {
            "$ref": "#/components/schemas/IdentityUserSchema"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ResponseModel[UserSchema]"
      },
      "IdentityResponseModel_UserWithToken_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "data": {
            "$ref": "#/components/schemas/IdentityUserWithToken"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ResponseModel[UserWithToken]"
      },
      "IdentityResponseModel_list_CompanySchema__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/IdentityCompanySchema"
            },
            "type": "array",
            "title": "Data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ResponseModel[list[CompanySchema]]"
      },
      "IdentityResponseModel_list_UserSchema__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/IdentityUserSchema"
            },
            "type": "array",
            "title": "Data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ResponseModel[list[UserSchema]]"
      },
      "IdentityUserLoginData": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "UserLoginData",
        "examples": [
          {
            "password": "mypassword",
            "username": "captain_zhang"
          }
        ]
      },
      "IdentityUserRegisterData": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "company_id": {
            "type": "integer",
            "title": "Company Id"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password",
          "phone",
          "company_id"
        ],
        "title": "UserRegisterData",
        "examples": [
          {
            "company_id": 1,
            "password": "mypassword",
            "phone": "13900000001",
            "username": "captain_zhang"
          }
        ]
      },
      "IdentityUserSchema": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "company_id": {
            "type": "integer",
            "title": "Company Id"
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin"
          },
          "is_system_admin": {
            "type": "boolean",
            "title": "Is System Admin"
          },
          "disabled": {
            "type": "boolean",
            "title": "Disabled"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "phone",
          "company_id",
          "is_admin",
          "is_system_admin",
          "disabled"
        ],
        "title": "UserSchema",
        "examples": [
          {
            "company_id": 1,
            "disabled": false,
            "id": 1,
            "is_admin": false,
            "is_system_admin": false,
            "phone": "13900000001",
            "username": "captain_zhang"
          }
        ]
      },
      "IdentityUserUpdate": {
        "properties": {
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "type": "object",
        "title": "UserUpdate",
        "examples": [
          {
            "phone": "13988888888"
          }
        ]
      },
      "IdentityUserWithToken": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "company_id": {
            "type": "integer",
            "title": "Company Id"
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin"
          },
          "is_system_admin": {
            "type": "boolean",
            "title": "Is System Admin"
          },
          "disabled": {
            "type": "boolean",
            "title": "Disabled"
          },
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "phone",
          "company_id",
          "is_admin",
          "is_system_admin",
          "disabled",
          "token"
        ],
        "title": "UserWithToken",
        "examples": [
          {
            "company_id": 1,
            "disabled": false,
            "id": 1,
            "is_admin": false,
            "is_system_admin": false,
            "phone": "13900000001",
            "username": "captain_zhang"
          }
        ]
      },
      "IdentityValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}